Enh: http_daemon: Check interrupted flag in own listening thread
Actually the http_daemon thread was ran as a "daemon-thread" :
once sys.exit() is called the thread is also exited.

but depending on the conditions the sys.exit() may NOT be called, or not directly,
leading to delayed / longer time for the daemon to exit.

It's better to pro-actively check our "interrupted" flag by/in this thread itself.

Also make sure to join the childs/clients threads at the end of the http_thread.